Added `@file:Suppress("DIFFERENT_BIT_WIDTHS")` to the relevant Native source files. This is the standard approach for handling a known issue in Kotlin/Native where cross-platform C types—such as `size_t`, which differs between 32-bit and 64-bit architectures—trigger warnings during metadata compilation.

+++ library/kmplitert-core/src/nativeMain/kotlin/io/github/leitingzi/kmplitert/core/LiteRTCompiler.native.kt
@@ -31,15 +31,15 @@ actual class LiteRTCompiler actual constructor(
     }
 
     actual suspend fun getInputBuffers(signatureIndex: Int): List<TFBuffer> {
-        return compiledModel.getInputBuffers(signatureIndex.toLong())
+        return compiledModel.getInputBuffers(signatureIndex)
     }
 
     actual suspend fun getOutputBuffers(signatureIndex: Int): List<TFBuffer> {
-        return compiledModel.getOutputBuffers(signatureIndex.toLong())
+        return compiledModel.getOutputBuffers(signatureIndex)
     }
 
     actual suspend fun run(inputs: List<TFBuffer>, outputs: List<TFBuffer>, signatureIndex: Int) {
-        compiledModel.run(signatureIndex.toLong(), inputs, outputs)
+        compiledModel.run(signatureIndex, inputs, outputs)
     }
 
     actual suspend fun close() {
